Refactor dependency injection, remove legacy user session references and drift db
