--- lib/core/database/drift/entities/point/point_geometry_table.dart
@@ -1,7 +1,0 @@
-import 'package:drift/drift.dart';
-
-class PointGeometryTable extends Table {
-  IntColumn get id => integer().autoIncrement()(); // Primary Key
-  TextColumn get type => text()(); // "type" field in PointGeometryDto
-  TextColumn get coordinates => text()(); // Coordinates stored as a JSON string
-}
