--- lib/core/database/drift/entities/track/track_table.dart
@@ -1,10 +1,0 @@
-import 'package:drift/drift.dart';
-
-class TrackTable extends Table {
-  IntColumn get id => integer().autoIncrement()();
-  TextColumn get trackId => text()();
-  DateTimeColumn get startTimestamp => dateTime()();
-  DateTimeColumn get endTimestamp => dateTime().nullable()();
-  BoolColumn get active => boolean().withDefault(const Constant(true))();
-  IntColumn get userId => integer()();
-}
