+++ lib/core/di/dependency_injection.dart
@@ -13,32 +13,28 @@ import 'package:dawarich/features/stats/application/services/stats_service.dart'
 import 'package:dawarich/features/tracking/application/services/system_settings_service.dart';
 import 'package:dawarich/features/tracking/application/services/track_service.dart';
 import 'package:dawarich/features/tracking/application/services/tracker_preferences_service.dart';
-import 'package:dawarich/core/session/application/user_session_service.dart';
 import 'package:dawarich/core/network/configs/api_config_manager.dart';
 import 'package:dawarich/core/network/dio_client.dart';
 import 'package:dawarich/features/auth/data/repositories/connect_repository.dart';
 import 'package:dawarich/features/tracking/data/repositories/hardware_repository.dart';
 import 'package:dawarich/core/network/repositories/api_point_repository.dart';
 import 'package:dawarich/features/stats/data/repositories/stats_repository.dart';
-import 'package:dawarich/core/database/%20repositories/objectbox/objectbox_point_local_repository.dart';
-import 'package:dawarich/core/database/%20repositories/objectbox/objectbox_track_repository.dart';
-import 'package:dawarich/core/database/%20repositories/objectbox/objectbox_user_storage_repository.dart';
+import 'package:dawarich/core/database/repositories/objectbox/objectbox_point_local_repository.dart';
+import 'package:dawarich/core/database/repositories/objectbox/objectbox_track_repository.dart';
+import 'package:dawarich/core/database/repositories/objectbox/objectbox_user_storage_repository.dart';
 import 'package:dawarich/features/tracking/data/repositories/tracker_preferences_repository.dart';
-import 'package:dawarich/core/session/data/user_session_repository.dart';
 import 'package:dawarich/features/tracking/data/sources/battery_data_client.dart';
 import 'package:dawarich/features/tracking/data/sources/device_data_client.dart';
 import 'package:dawarich/features/tracking/data/sources/gps_data_client.dart';
 import 'package:dawarich/features/tracking/data/sources/connectivity_data_client.dart';
-import 'package:dawarich/core/database/drift/database/sqlite_client.dart';
 import 'package:dawarich/core/network/configs/api_config_manager_interfaces.dart';
 import 'package:dawarich/features/auth/data_contracts/interfaces/connect_repository_interfaces.dart';
 import 'package:dawarich/features/tracking/data_contracts/interfaces/hardware_repository_interfaces.dart';
-import 'package:dawarich/core/database/%20repositories/local_point_repository_interfaces.dart';
+import 'package:dawarich/core/database/repositories/local_point_repository_interfaces.dart';
 import 'package:dawarich/core/network/repositories/api_point_repository_interfaces.dart';
 import 'package:dawarich/features/stats/data_contracts/interfaces/stats_repository_interfaces.dart';
 import 'package:dawarich/features/tracking/data_contracts/interfaces/i_track_repository.dart';
 import 'package:dawarich/features/tracking/data_contracts/interfaces/tracker_preferences_repository_interfaces.dart';
-import 'package:dawarich/core/session/domain/legacy_user_session_repository_interfaces.dart';
 import 'package:dawarich/features/auth/data_contracts/interfaces/user_storage_repository_interfaces.dart';
 import 'package:dawarich/objectbox.g.dart';
 import 'package:dawarich/features/batch/presentation/models/batch_explorer_viewmodel.dart';
@@ -87,7 +83,6 @@ final class DependencyInjection {
       ]),
       dependsOn: [IApiConfigManager],
     );
-    getIt.registerLazySingleton<SQLiteClient>(() => SQLiteClient());
     getIt.registerSingletonAsync<Store>(() async {
       final dir = await getApplicationDocumentsDirectory();
       return openStore(directory: '${dir.path}/objectbox');
@@ -99,8 +94,6 @@ final class DependencyInjection {
         () => ConnectivityDataClient());
 
     // Repositories
-    getIt.registerLazySingleton<ILegacyUserSessionRepository>(
-        () => LegacyUserSessionRepository());
     getIt.registerLazySingleton<IUserStorageRepository>(
         () => ObjectBoxUserStorageRepository(getIt<Store>()));
     getIt.registerLazySingleton<IHardwareRepository>(() => HardwareRepository(
@@ -123,10 +116,8 @@ final class DependencyInjection {
 
     // Services
     getIt.registerSingletonWithDependencies<MigrationService>(
-        () => MigrationService(getIt<SQLiteClient>(), getIt<Store>()),
+        () => MigrationService(getIt<Store>()),
         dependsOn: [Store]);
-    getIt.registerLazySingleton<LegacyUserSessionService>(
-        () => LegacyUserSessionService(getIt<ILegacyUserSessionRepository>()));
     getIt.registerLazySingleton<SystemSettingsService>(
         () => SystemSettingsService());
     getIt.registerLazySingleton<ApiConfigService>(
