--- lib/core/session/domain/legacy_user_session_repository_interfaces.dart
@@ -1,5 +1,0 @@
-abstract interface class ILegacyUserSessionRepository {
-  Future<int> getCurrentUserId();
-  Future<void> setCurrentUserId(int userId);
-  Future<void> clearSession();
-}
