+++ lib/features/auth/data_contracts/data_transfer_objects/users/user_settings_dto.dart
@@ -1,4 +1,3 @@
-import 'package:dawarich/core/database/drift/database/sqlite_client.dart';
 
 class UserSettingsDto {
   String immichUrl;
@@ -20,13 +19,6 @@ class UserSettingsDto {
         photoprismApiKey: json["photoprism_api_key"]);
   }
 
-  factory UserSettingsDto.fromDatabase(UserSettingsTableData userSettings) {
-    return UserSettingsDto(
-        immichUrl: userSettings.immichUrl,
-        immichApiKey: userSettings.immichApiKey,
-        photoprismUrl: userSettings.photoprismUrl,
-        photoprismApiKey: userSettings.photoprismApiKey);
-  }
 
   // Map<String, dynamic> toJson() {
   //   return {
