+++ lib/features/migration/presentation/models/migration_viewmodel.dart
@@ -1,5 +1,4 @@
 import 'package:dawarich/features/migration/application/services/migration_service.dart';
-import 'package:dawarich/core/session/application/user_session_service.dart';
 import 'package:dawarich/core/di/dependency_injection.dart';
 import 'package:dawarich/core/routing/app_router.dart';
 import 'package:flutter/cupertino.dart';
